+++ system-tests/projects/e2e/cypress/plugins/index.js
@@ -4,7 +4,7 @@ const _ = require('lodash')
 const Jimp = require('jimp')
 const path = require('path')
 const Promise = require('bluebird')
-const { useFixedBrowserLaunchSize } = require('../../../utils')
+const { useFixedBrowserLaunchSize } = require('@tooling/system-tests/lib/pluginUtils')
 
 /**
  * @type {Cypress.PluginConfig}